⚡️ repositoryクラスの共通の処理を継承によって実装
- connect()
- disconnect()
- request()
